fix: fix CP of uncompressed size annotation add
The patch from upstream conflicted with the blocksize changes because
the upstream patch changes Compress to have a pointer receiver so we can
store the number of bytes read, and the blocksize changes added a few
uses of Compressor types as a non-pointer type that can't implement the
Compressor interface if it has a pointer receiver.

The fix is to just always use pointer types.

Signed-off-by: Michael McCracken <mikmccra@cisco.com>