fix: convert overlay whiteout char dev into userxattr
When extracting a OCI tar layer, a whiteout entry a/.wh.b will be
converted to "mknod a/b c 0 0".  However, if in the same layer, we also
create a a/b/c entry, it will fail since /a/b is a char dev. In this
situation, we remove /a/b and instead set user.overlay.opaque on "a/b".

Furthermore, overlayfs creates an opaque entry when a new dir entry is
created. We want to avoid that and only emit a whiteout if underlying
layers already had the same dir entry.

Signed-off-by: Ramkumar Chinchani <rchincha@cisco.com>